modified userinfo

+++ schedule/views.py
@@ -4,9 +4,11 @@ from django.http import HttpResponseRedirect
 from .forms import *
 from .models import *
 from datetime import datetime
+from django.contrib import messages
 
 def events(request):
-	return render(request,'schedule/events.html')
+	events = Event.objects.all()
+	return render(request,'schedule/events.html',{'events' : events})
 
 def update_event(request):
 	if request.method == 'POST':
@@ -14,9 +16,12 @@ def update_event(request):
 		form = update_detail(request.POST)
 		if form.is_valid():
 			user = request.user
-			query = Event.objects.get(name = name)
-			form1  = update_detail(request.POST, instance = query)
-			form1.save()
+			if Event.objects.filter(name = name).exists():
+				query = Event.objects.get(name = name)
+				form1  = update_detail(request.POST, instance = query)
+				form1.save()
+			else :
+				messages.error(request,('Enter a valid event name'))
 		return HttpResponseRedirect('')
 	else:
 		form = update_detail()
@@ -29,9 +34,12 @@ def delete_event(request):
 		form = delete_details(request.POST)
 		if form.is_valid():
 			user = request.user
-			query = Event.objects.get(name = name)
-			query.delete()
-		return HttpResponseRedirect('')
+			if Event.objects.filter(name = name).exists():
+				query = Event.objects.get(name = name)
+				query.delete()
+			else :
+				messages.error(request,('Enter a valid event name'))
+			return HttpResponseRedirect('')
 	else:
 		form = delete_details()
 		events = Event.objects.all()
@@ -43,19 +51,17 @@ def create_event(request):
 		if form.is_valid():
 			user = request.user
 			name = request.POST.get('name')
-			date = request.POST.get('date')
-			start_time = request.POST.get('start_time')
-			end_time = request.POST.get('end_time')
-			description = request.POST.get('description')
-			event_obj = Event(user=user, name = name, date = date, start_time = start_time, end_time = end_time, description = description,)
-			event_obj.save()
-
-			return HttpResponseRedirect('')
+			if not Event.objects.filter(name = name).exists():
+				date = request.POST.get('date')
+				start_time = request.POST.get('start_time')
+				end_time = request.POST.get('end_time')
+				description = request.POST.get('description')
+				event_obj = Event(user=user, name = name, date = date, start_time = start_time, end_time = end_time, description = description,)
+				event_obj.save()
+			else:
+				messages.error(request,('An event with that name already exists'))
+				return HttpResponseRedirect('')
+			return render(request,'schedule/events.html')
 	else:
 		form = event_details()
-
 	return render(request, 'schedule/new_event.html',{'form': form,})
-
-def all_events(request):
-	events = Event.objects.all()
-	return render(request,'schedule/All_events.html',{'events' : events})
